worlds -> instances

+++ mcadmin/libraries/mc_server/backup.py
@@ -2,6 +2,8 @@ import logging
 import os
 import asyncio
 import shutil
+import json
+import aiofiles
 
 
 __all__ = [
@@ -17,9 +19,13 @@ class McServerBackupError(Exception):
 
 
 class McServerBackup:
-    """Low level Minecraft world backup manager"""
+    """Low level Minecraft server backup manager"""
 
-    backup_dirs: list[str] = ["world", "mods"]
+    backup_targets: list[tuple[str, str]] = [
+        ("world", "dir"),
+        ("mods", "dir"),
+        ("server_info.json", "file"),
+    ]
 
     def __init__(self, instance_dir: str, backups_dir: str) -> None:
         self._instance_dir: str = instance_dir
@@ -33,15 +39,23 @@ class McServerBackup:
             logger.info(f"Creating backup directory {backup_dir}")
             os.makedirs(backup_dir)
 
-        for d in self.backup_dirs:
-            src_dir = os.path.join(self._instance_dir, d)
-            dst_dir = os.path.join(backup_dir, d)
+        for d, t in self.backup_targets:
+            src_path = os.path.join(self._instance_dir, d)
+            dst_path = os.path.join(backup_dir, d)
 
-            if os.path.exists(src_dir):
-                await asyncio.to_thread(shutil.copytree, src_dir, dst_dir)
+            if not os.path.exists(src_path):
+                if t == "dir":
+                    os.makedirs(dst_path)
+                else:
+                    async with aiofiles.open(dst_path, "w") as f:
+                        await f.write("")
+
+                continue
+
+            if t == "dir":
+                await asyncio.to_thread(shutil.copytree, src_path, dst_path)
             else:
-                # create empty directory
-                os.makedirs(dst_dir)
+                await asyncio.to_thread(shutil.copy2, src_path, dst_path)
 
         logger.info(f"Successfully backed up data to {backup}")
 
@@ -57,9 +71,15 @@ class McServerBackup:
             dst_dir = os.path.join(self._instance_dir, d)
 
             if os.path.exists(dst_dir):
-                await asyncio.to_thread(shutil.rmtree, dst_dir)
+                if os.path.isdir(dst_dir):
+                    await asyncio.to_thread(shutil.rmtree, dst_dir)
+                else:
+                    await asyncio.to_thread(os.remove, dst_dir)
 
-            await asyncio.to_thread(shutil.copytree, src_dir, dst_dir)
+            if os.path.isdir(src_dir):
+                await asyncio.to_thread(shutil.copytree, src_dir, dst_dir)
+            else:
+                await asyncio.to_thread(shutil.copy2, src_dir, dst_dir)
 
     async def delete_backup(self, backup: str) -> None:
         """Delete a backup with the given name"""
